Unmangle symbols when enumerating and matching lookup symbol.
Fixes issue 581.

Most users ignore argument overloads when doing symbol lookup, so we
support that by matching either the whole unmangled symbol name, or an
exact prefix match followed by an open paren.

Instead of using libelftc's unmangling facilities, we chose to use
demangle.cc from google-glog because it was about 8x faster than
libelftc.

SVN-Revision: 1078

+++ ext/drsyms/demangle.cc
@@ -29,6 +29,8 @@
 //
 // Author: Satoru Takabayashi
 //
+// Retreived from http://code.google.com/p/google-glog/ at r91.
+//
 // For reference check out:
 // http://www.codesourcery.com/public/cxx-abi/abi.html#mangling
 //
@@ -397,7 +399,7 @@ static void MaybeCancelLastSeparator(State *state) {
 // namespace.
 static bool IdentifierIsAnonymousNamespace(State *state) {
   const char anon_prefix[] = "_GLOBAL__N_";
-  return (state->number > sizeof(anon_prefix) - 1 &&  // Should be longer.
+  return (state->number > (int)sizeof(anon_prefix) - 1 &&  // Should be longer.
           StrPrefix(state->mangled_cur, anon_prefix));
 }
